[POST-RELEASE] Spawnable AI Tests (NOT WORKING)
Tried to get the slottable AI system to instantiate well, however for some reason the >>instantiated<< scriptable objects inherit all of their behaviours from a single instance of their mother script. Haven't figured out how to fix this yet.

AI is broken in this patch so revert to another repo if no fixed.